apps/app3/pyproject.toml
[project]
name = "app3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.0.3",
    "pydantic-settings>=2.12.0",
    "langchain-core>=0.3.6",
]
